+++ src/collective/classification/folder/content/classification_folder.py
@@ -36,16 +36,6 @@ from zope.interface import invariant
 from zope.lifecycleevent import IObjectRemovedEvent
 from zope.schema.fieldproperty import FieldProperty
 
-import pkg_resources
-
-
-try:
-    pkg_resources.get_distribution("collective.querynextprev")
-except pkg_resources.DistributionNotFound:
-    HAS_QUERYNEXTPREV = False
-else:
-    HAS_QUERYNEXTPREV = True
-
 
 class IClassificationFolder(model.Schema):
     """Marker interface and Dexterity Python Schema for ClassificationFolder"""
@@ -148,7 +138,7 @@ class ClassificationFolder(Container):
 
 
 def on_create(obj, event):
-    """Configures faceted navigation. Disables querynextprev. Increases settings counter."""
+    """Configures faceted navigation. Increases settings counter."""
     notify(FacetedWillBeEnabledEvent(obj))
     alsoProvides(obj, IClassificationFacetedNavigable)
     if not IDisableSmartFacets.providedBy(obj):
@@ -156,11 +146,6 @@ def on_create(obj, event):
     if not IHidePloneRightColumn.providedBy(obj):
         alsoProvides(obj, IHidePloneRightColumn)
     notify(FacetedEnabledEvent(obj))
-
-    if HAS_QUERYNEXTPREV:
-        from collective.querynextprev.interfaces import INextPrevNotNavigable
-
-        alsoProvides(obj, INextPrevNotNavigable)
 
     IFacetedLayout(obj).update_layout("folder-listing-view")